Refactoring project structure and information
