+++ src/pages/404.js
@@ -1,11 +1,10 @@
 import React from 'react'
-
-import Layout from '../components/layout'
-import SEO from '../components/seo'
+import Layout from '../components/HOC/Layout'
+import SEO from '../components/Utility/SEO'
 
 const NotFoundPage = () => (
   <Layout>
-    <SEO title="404: Not found" />
+    <SEO title='404: Not found' />
     <h1>NOT FOUND</h1>
     <p>You just hit a route that doesn&#39;t exist... the sadness.</p>
   </Layout>
