+++ src/pages/index.js
@@ -1,20 +1,11 @@
 import React from 'react'
-import { Link } from 'gatsby'
-
-import Layout from '../components/layout'
-import Image from '../components/image'
-import SEO from '../components/seo'
+import Layout from '../components/HOC/Layout';
+import SEO from '../components/Utility/SEO';
 
 const IndexPage = () => (
   <Layout>
     <SEO title="Home" keywords={[`gatsby`, `application`, `react`]} />
-    <h1>Hi people</h1>
-    <p>Welcome to your new Gatsby site.</p>
-    <p>Now go build something great.</p>
-    <div style={{ maxWidth: `300px`, marginBottom: `1.45rem` }}>
-      <Image />
-    </div>
-    <Link to="/page-2/">Go to page 2</Link>
+    <h1>Welcome To My Gatsby Blog!</h1>
   </Layout>
 )
